scripts/parser/requirements.txt
requests==2.28.1
beautifulsoup4==4.11.1
psycopg2-binary==2.9.5